libpod: Move part of (*Container).stat to container_stat_linux.go
The logic that treats running containers differently from stopped
containers is not needed on FreeBSD where the container mounts live in
a global mount namespace.

[NO NEW TESTS NEEDED]
